fix: Apply pre-commit formatting fixes
