+++ playlist_etl/main.py
@@ -1,12 +1,12 @@
 import os
 
-from playlist_etl.aggregate2 import Aggregate
+from playlist_etl.aggregate import Aggregate
 from playlist_etl.config import ISRC_CACHE_COLLECTION, YOUTUBE_URL_CACHE_COLLECTION
 from playlist_etl.helpers import set_secrets
 from playlist_etl.services import AppleMusicService, SpotifyService, YouTubeService
-from playlist_etl.transform2 import Transform
+from playlist_etl.transform import Transform
 from playlist_etl.utils import CacheManager, MongoDBClient, WebDriverManager
-from playlist_etl.view_count2 import ViewCountTrackProcessor
+from playlist_etl.view_count import ViewCountTrackProcessor
 
 
 def transform(
@@ -45,9 +45,7 @@ def main() -> None:
         api_key=os.getenv("GOOGLE_API_KEY"),
         cache_manager=CacheManager(mongo_client, YOUTUBE_URL_CACHE_COLLECTION),
     )
-    apple_music_service = AppleMusicService(
-        CacheManager(mongo_client, YOUTUBE_URL_CACHE_COLLECTION)
-    )
+    apple_music_service = AppleMusicService(CacheManager(mongo_client, YOUTUBE_URL_CACHE_COLLECTION))
 
     transform(mongo_client, spotify_service, youtube_service, apple_music_service)
     aggregate(mongo_client)
