+++ playlist_etl/transform.py
@@ -1,330 +1,258 @@
 import concurrent.futures
-import os
-import re
-from urllib.parse import unquote
+import json
+from collections import defaultdict
 
-import requests
-from bs4 import BeautifulSoup
-
-from playlist_etl.extract import PLAYLIST_GENRES, SERVICE_CONFIGS
-from playlist_etl.helpers import get_logger, set_secrets
-from playlist_etl.utils import (
-    clear_collection,
-    get_mongo_client,
-    get_spotify_client,
-    insert_or_update_data_to_mongo,
-    read_cache_from_mongo,
-    read_data_from_mongo,
-    update_cache_in_mongo,
+from playlist_etl.config import (
+    RAW_PLAYLISTS_COLLECTION,
+    TRACK_COLLECTION,
+    TRACK_PLAYLIST_COLLECTION,
 )
+from playlist_etl.helpers import get_logger
+from playlist_etl.models import (
+    GenreName,
+    Playlist,
+    Track,
+    TrackRank,
+    TrackSourceServiceName,
+)
+from playlist_etl.mongo_db_client import MongoDBClient
+from playlist_etl.services import AppleMusicService, SpotifyService, YouTubeService
 
-RAW_PLAYLISTS_COLLECTION = "raw_playlists"
-TRANSFORMED_DATA_COLLECTION = "transformed_playlists"
-YOUTUBE_CACHE_COLLECTION = "youtube_cache"
-ISRC_CACHE_COLLECTION = "isrc_cache"
-APPLE_MUSIC_ALBUM_COVER_CACHE_COLLECTION = "apple_music_album_cover_cache"
-
-MAX_THREADS = 50
+MAX_THREADS = 100
 
 logger = get_logger(__name__)
 
 
-class Track:
+class Transform:
+    """
+    Converts raw tracks to Track Object
+    Converts raw playlists to Playlist Object.
+    """
+
     def __init__(
         self,
-        track_name,
-        artist_name,
-        track_url,
-        rank,
-        source_name,
-        genre_name,
-        album_cover_url=None,
-        isrc=None,
+        mongo_client: MongoDBClient,
+        spotify_service: SpotifyService,
+        youtube_service: YouTubeService,
+        apple_music_service: AppleMusicService,
     ):
-        self.isrc = isrc
-        self.track_name = track_name
-        self.artist_name = artist_name
-        self.track_url = track_url
-        self.album_cover_url = album_cover_url
-        self.rank = rank
-        self.source_name = source_name
-        self.genre_name = genre_name
-        self.youtube_url = None
+        self.mongo_client = mongo_client
+        self.spotify_service = spotify_service
+        self.youtube_service = youtube_service
+        self.apple_music_service = apple_music_service
+        self.tracks: dict[str, Track] = {}
+        self.playlist_ranks: defaultdict[tuple[str, str], list[TrackRank]] = defaultdict(list)
 
-    def to_dict(self):
-        return self.__dict__
+    def transform(self) -> None:
+        logger.info("Starting the transformation process")
+        raw_playlists = self.mongo_client.get_collection(RAW_PLAYLISTS_COLLECTION)
+        self.build_track_objects(raw_playlists)
+        self.set_youtube_urls()
+        self.set_apple_music_album_covers()
+        self.set_spotify_urls()
+        self.overwrite()
+        logger.info("Transformation process completed successfully")
 
-    @staticmethod
-    def from_dict(data):
-        return Track(**data)
+    def build_track_objects(self, raw_playlists: dict) -> None:
+        logger.info("Reading raw playlists from MongoDB")
 
-    def set_isrc(self, spotify_client, mongo_client):
-        if not self.isrc:
-            self.isrc = get_isrc_from_spotify_api(self.track_name, self.artist_name, spotify_client, mongo_client)
+        for genre_name in GenreName:
+            for service_name in TrackSourceServiceName:
+                genre_name_value = genre_name.value
+                service_name_value = service_name.value
 
-    def set_youtube_url(self, mongo_client):
-        if not self.youtube_url:
-            self.youtube_url = get_youtube_url_by_track_and_artist_name(self.track_name, self.artist_name, mongo_client)
+                raw_playlist = raw_playlists.find_one(
+                    {"genre_name": genre_name_value, "service_name": service_name_value}
+                )
+                logger.info(f"Processing playlist for genre {genre_name_value} from {service_name_value}")
+                if not raw_playlist:
+                    logger.warning(f"No raw playlist found for {service_name_value} and genre {genre_name_value}")
+                    raise ValueError("No raw playlist found")
+                data = json.loads(raw_playlist["data_json"])
+                self.convert_to_track_objects(data, service_name_value, genre_name_value)
 
-    def set_apple_music_album_cover_url(self, mongo_client):
-        if not self.album_cover_url:
-            self.album_cover_url = get_apple_music_album_cover(self.track_url, mongo_client)
+    def format_tracks(self) -> list[dict]:
+        formatted_tracks = [
+            {"isrc": isrc, "track_data": track.model_dump(exclude={"isrc"})}
+            for isrc, track in self.tracks.items()
+            if track.isrc is not None
+        ]
+        return formatted_tracks
 
+    def format_playlist_ranks(self) -> list[dict]:
+        formatted_ranks = [
+            Playlist(
+                service_name=TrackSourceServiceName(service_name),
+                genre_name=genre_name,
+                tracks=[TrackRank(isrc=rank.isrc, rank=rank.rank, sources=rank.sources) for rank in ranks],
+            ).model_dump()
+            for (service_name, genre_name), ranks in self.playlist_ranks.items()
+        ]
+        return formatted_ranks
 
-def convert_to_track_objects(data, service_name, genre_name):
-    logger.info(f"Converting data to track objects for {service_name} and genre {genre_name}")
-    match service_name:
-        case "AppleMusic":
-            return convert_apple_music_raw_export_to_track_type(data, genre_name)
-        case "Spotify":
-            return convert_spotify_raw_export_to_track_type(data, genre_name)
-        case "SoundCloud":
-            return convert_soundcloud_raw_export_to_track_type(data, genre_name)
-        case _:
+    def convert_to_track_objects(self, data: dict, service_name: str, genre_name: str) -> None:
+        logger.info(f"Converting data to track objects for {service_name} and genre {genre_name}")
+        if service_name == TrackSourceServiceName.APPLE_MUSIC:
+            self.convert_apple_music_raw_export_to_track_type(data, genre_name)
+        elif service_name == TrackSourceServiceName.SPOTIFY:
+            self.convert_spotify_raw_export_to_track_type(data, genre_name)
+        elif service_name == TrackSourceServiceName.SOUNDCLOUD:
+            self.convert_soundcloud_raw_export_to_track_type(data, genre_name)
+        else:
             raise ValueError("Unknown service name")
 
+    def get_track(self, isrc: str) -> Track:
+        if isrc in self.tracks:
+            track = self.tracks[isrc]
+        else:
+            track = Track(isrc=isrc)
+            self.tracks[isrc] = track
+        return track
 
-def convert_apple_music_raw_export_to_track_type(data, genre_name):
-    logger.info(f"Converting Apple Music data for genre {genre_name}")
-    tracks = []
-    for key, track_data in data["album_details"].items():
-        if key.isdigit():
-            rank = int(key) + 1
-            track_name = track_data["name"]
-            artist_name = track_data["artist"]
-            track_url = track_data["link"]
-            track = Track(track_name, artist_name, track_url, rank, "AppleMusic", genre_name)
-            tracks.append(track)
-    return tracks
+    def convert_apple_music_raw_export_to_track_type(self, data: dict, genre_name: str) -> None:
+        logger.info(f"Converting Apple Music data for genre {genre_name}")
+        # Handle case where data might be a JSON string instead of parsed dict
+        if isinstance(data, str):
+            data = json.loads(data)
+        for key, track_data in data["album_details"].items():
+            if key.isdigit():
+                track_name = track_data["name"]
+                artist_name = track_data["artist"]
+                isrc = self.spotify_service.get_isrc(track_name, artist_name)
 
+                if isrc is None:
+                    logger.warning(f"ISRC not found for Apple Music track: {track_name} by {artist_name}")
+                    continue
 
-def convert_soundcloud_raw_export_to_track_type(data, genre_name):
-    logger.info(f"Converting SoundCloud data for genre {genre_name}")
-    tracks = []
-    for i, item in enumerate(data["tracks"]["items"]):
-        track_name = item["title"]
-        track_url = item["permalink"]
-        rank = i + 1
-        artist_name = item["user"]["name"]
-        album_cover_url = item["artworkUrl"]
-        isrc = item["publisher"]["isrc"]
+                track = self.get_track(isrc)
+                track.apple_music_track_data.track_name = track_name
+                track.apple_music_track_data.artist_name = artist_name
+                track.apple_music_track_data.track_url = track_data["link"]
 
-        if " - " in track_name:
-            artist_name, track_name = track_name.split(" - ", 1)
-        track = Track(
-            track_name,
-            artist_name,
-            track_url,
-            rank,
-            "SoundCloud",
-            genre_name,
-            album_cover_url,
-            isrc,
-        )
-        tracks.append(track)
-    return tracks
+                self.playlist_ranks[(TrackSourceServiceName.APPLE_MUSIC.value, genre_name)].append(
+                    TrackRank(
+                        isrc=isrc,
+                        rank=int(key) + 1,
+                        sources={TrackSourceServiceName.APPLE_MUSIC: int(key) + 1},
+                    )
+                )
 
+    def convert_soundcloud_raw_export_to_track_type(self, data: dict, genre_name: str) -> None:
+        logger.info(f"Converting SoundCloud data for genre {genre_name}")
+        # Handle case where data might be a JSON string instead of parsed dict
+        if isinstance(data, str):
+            data = json.loads(data)
+        for i, item in enumerate(data["tracks"]["items"]):
+            isrc = item["publisher"]["isrc"]
+            if not isrc:
+                continue
 
-def convert_spotify_raw_export_to_track_type(data, genre_name):
-    logger.info(f"Converting Spotify data for genre {genre_name}")
-    tracks = []
-    for rank, item in enumerate(data["items"]):
-        track_info = item["track"]
+            track_name = item["title"]
+            artist_name = item["user"]["name"]
 
-        if not track_info:
-            continue
+            if " - " in track_name:
+                artist_name, track_name = track_name.split(" - ", 1)
 
-        isrc = track_info["external_ids"]["isrc"]
-        track_name = track_info["name"]
-        track_url = track_info["external_urls"]["spotify"]
-        artist_name = ", ".join(artist["name"] for artist in track_info["artists"])
-        album_cover_url = track_info["album"]["images"][0]["url"]
-        track = Track(
-            track_name,
-            artist_name,
-            track_url,
-            rank + 1,
-            "Spotify",
-            genre_name,
-            album_cover_url,
-            isrc,
-        )
-        tracks.append(track)
-    return tracks
+            track = self.get_track(isrc)
+            track.soundcloud_track_data.track_name = track_name
+            track.soundcloud_track_data.artist_name = artist_name
+            track.soundcloud_track_data.track_url = item["permalink"]
+            track.soundcloud_track_data.album_cover_url = item["artworkUrl"]
 
+            self.playlist_ranks[(TrackSourceServiceName.SOUNDCLOUD.value, genre_name)].append(
+                TrackRank(
+                    isrc=isrc,
+                    rank=i + 1,
+                    sources={TrackSourceServiceName.SOUNDCLOUD: i + 1},
+                )
+            )
 
-def get_isrc_from_spotify_api(track_name, artist_name, spotify_client, mongo_client):
-    cache_key = f"{track_name}|{artist_name}"
-    isrc_cache = read_cache_from_mongo(mongo_client, ISRC_CACHE_COLLECTION)
+    def convert_spotify_raw_export_to_track_type(self, data: dict, genre_name: str) -> None:
+        logger.info(f"Converting Spotify data for genre {genre_name}")
+        # Handle case where data might be a JSON string instead of parsed dict
+        if isinstance(data, str):
+            data = json.loads(data)
+        for i, item in enumerate(data["items"]):
+            track_info = item["track"]
+            if not track_info:
+                continue
+            isrc = track_info["external_ids"]["isrc"]
 
-    if cache_key in isrc_cache:
-        logger.info(f"Cache hit for ISRC: {cache_key}")
-        return isrc_cache[cache_key]
+            track = self.get_track(isrc)
+            track.spotify_track_data.track_name = track_info["name"]
+            track.spotify_track_data.artist_name = ", ".join(artist["name"] for artist in track_info["artists"])
+            track.spotify_track_data.track_url = track_info["external_urls"]["spotify"]
+            track.spotify_track_data.album_cover_url = track_info["album"]["images"][0]["url"]
 
-    logger.info(f"ISRC Spotify Lookup Cache miss for {track_name} by {artist_name}")
+            self.playlist_ranks[(TrackSourceServiceName.SPOTIFY.value, genre_name)].append(
+                TrackRank(
+                    isrc=isrc,
+                    rank=i + 1,
+                    sources={TrackSourceServiceName.SPOTIFY: i + 1},
+                )
+            )
 
-    def search_spotify(query):
-        try:
-            results = spotify_client.search(q=query, type="track", limit=1)
-            tracks = results["tracks"]["items"]
-            if tracks:
-                return tracks[0]["external_ids"]["isrc"]
-            return None
-        except Exception as e:
-            logger.info(f"Error searching Spotify with query '{query}': {e}")
-            return None
-
-    track_name_no_parens = re.sub(r"\([^()]*\)", "", track_name.lower())
-    queries = [
-        f"track:{track_name_no_parens} artist:{artist_name}",
-        f"{track_name_no_parens} {artist_name}",
-        f"track:{track_name.lower()} artist:{artist_name}",
-    ]
-
-    for query in queries:
-        isrc = search_spotify(query)
-        if isrc:
-            logger.info(f"Found ISRC for {track_name} by {artist_name}: {isrc}")
-            update_cache_in_mongo(mongo_client, ISRC_CACHE_COLLECTION, cache_key, isrc)
-            return isrc
-
-    logger.info(f"No track found on Spotify using queries: {queries} for {track_name} by {artist_name}")
-    return None
-
-
-def get_youtube_url_by_track_and_artist_name(track_name, artist_name, mongo_client):
-    cache_key = f"{track_name}|{artist_name}"
-    youtube_cache = read_cache_from_mongo(mongo_client, YOUTUBE_CACHE_COLLECTION)
-
-    if cache_key in youtube_cache:
-        return youtube_cache[cache_key]
-
-    query = f"{track_name} {artist_name}"
-    api_key = os.getenv("GOOGLE_API_KEY")
-
-    youtube_search_url = f"https://www.googleapis.com/youtube/v3/search?part=snippet&q={query}&key={api_key}"
-    response = requests.get(youtube_search_url)
-
-    if response.status_code == 200:
-        data = response.json()
-        if data["items"]:
-            video_id = data["items"][0]["id"].get("videoId")
-            if video_id:
-                youtube_url = f"https://www.youtube.com/watch?v={video_id}"
-                logger.info(f"Updating YouTube cache for {track_name} by {artist_name} with URL {youtube_url}")
-                update_cache_in_mongo(mongo_client, YOUTUBE_CACHE_COLLECTION, cache_key, youtube_url)
-                return youtube_url
-        logger.info(f"No video found for {track_name} by {artist_name}")
-        return None
-    else:
-        logger.info(f"Error: {response.status_code}, {response.text}")
-        if response.status_code == 403 and "quotaExceeded" in response.text:
-            raise ValueError(f"Could not get Youtube URL for {track_name} {artist_name}")
-        return None
-
-
-def get_apple_music_album_cover(url_link, mongo_client):
-    cache_key = url_link
-    album_cover_cache = read_cache_from_mongo(mongo_client, APPLE_MUSIC_ALBUM_COVER_CACHE_COLLECTION)
-
-    if cache_key in album_cover_cache:
-        return album_cover_cache[cache_key]
-
-    logger.info(f"Apple Music Album Cover Cache miss for URL: {url_link}")
-    response = requests.get(url_link, timeout=30)
-    response.raise_for_status()
-    doc = BeautifulSoup(response.text, "html.parser")
-
-    source_tag = doc.find("source", attrs={"type": "image/jpeg"})
-    if not source_tag or not source_tag.has_attr("srcset"):
-        raise ValueError("Album cover URL not found")
-
-    srcset = source_tag["srcset"]
-    album_cover_url = unquote(srcset.split()[0])
-    update_cache_in_mongo(
-        mongo_client,
-        APPLE_MUSIC_ALBUM_COVER_CACHE_COLLECTION,
-        cache_key,
-        album_cover_url,
-    )
-    return album_cover_url
-
-
-def transform_playlists(mongo_client):
-    logger.info("Reading raw playlists from MongoDB")
-    raw_playlists = read_data_from_mongo(mongo_client, RAW_PLAYLISTS_COLLECTION)
-    logger.info(f"Raw playlists from MongoDB: {len(raw_playlists)} documents found")
-
-    spotify_client = get_spotify_client()
-    for genre in PLAYLIST_GENRES:
-        logger.info(f"Processing genre: {genre}")
-        all_tracks = []
-        genre_playlists = [doc for doc in raw_playlists if doc["genre_name"] == genre]
-        logger.info(f"Found {len(genre_playlists)} playlists for genre {genre}")
-
-        for document in genre_playlists:
-            service_name = document["service_name"]
-            data = document["data_json"]
-            logger.info(f"Processing {service_name} playlist for genre {genre}")
-            tracks = convert_to_track_objects(data, service_name, genre)
-            logger.info(f"Converted {len(tracks)} tracks for {service_name} in genre {genre}")
-            all_tracks.extend(tracks)
-
-        if not all_tracks:
-            logger.info(f"No tracks found for genre {genre}")
-            continue
-
-        logger.info("Setting ISRCs for all tracks")
-        with concurrent.futures.ThreadPoolExecutor(max_workers=MAX_THREADS) as executor:
-            futures = [executor.submit(track.set_isrc, spotify_client, mongo_client) for track in all_tracks]
-            concurrent.futures.wait(futures)
-
+    def set_youtube_urls(self) -> None:
         logger.info("Setting YouTube URLs for all tracks")
-        with concurrent.futures.ThreadPoolExecutor(max_workers=1) as executor:
-            futures = [executor.submit(track.set_youtube_url, mongo_client) for track in all_tracks]
-            concurrent.futures.wait(futures)
+        with concurrent.futures.ThreadPoolExecutor(max_workers=MAX_THREADS) as executor:
+            futures = [executor.submit(self.youtube_service.set_track_url, track) for track in self.tracks.values()]
+            for future in concurrent.futures.as_completed(futures):
+                future.result()
 
-        logger.info("Setting Apple Music album cover URLs for applicable tracks")
+    def set_apple_music_album_covers(self) -> None:
+        logger.info("Setting Apple Music album cover URLs for all tracks")
         with concurrent.futures.ThreadPoolExecutor(max_workers=MAX_THREADS) as executor:
             futures = [
-                executor.submit(track.set_apple_music_album_cover_url, mongo_client)
-                for track in all_tracks
-                if track.source_name == "AppleMusic"
+                executor.submit(self.set_apple_music_album_cover_url, track)
+                for track in self.tracks.values()
+                if track.apple_music_track_data
             ]
-            concurrent.futures.wait(futures)
+            for future in concurrent.futures.as_completed(futures):
+                future.result()
 
-        logger.info(f"Processing tracks by source and genre for {genre}")
-        for source in SERVICE_CONFIGS:
-            process_tracks(all_tracks, source, genre, mongo_client)
+    def set_apple_music_album_cover_url(self, track: Track) -> None:
+        if track.apple_music_track_data.track_name:
+            album_cover_url = self.apple_music_service.get_album_cover_url(track.apple_music_track_data.track_url)
+            track.apple_music_track_data.album_cover_url = album_cover_url
 
+    def set_spotify_urls(self) -> None:
+        logger.info("Setting Spotify URLs for all tracks")
+        with concurrent.futures.ThreadPoolExecutor(max_workers=MAX_THREADS) as executor:
+            futures = [
+                executor.submit(self.spotify_service.set_track_url, track)
+                for track in self.tracks.values()
+                if not track.spotify_track_data.track_url
+            ]
+            for future in concurrent.futures.as_completed(futures):
+                future.result()
 
-def process_tracks(tracks, source_name, genre, mongo_client):
-    logger.info(f"Processing {source_name} tracks for genre {genre}")
-    filtered_tracks = [
-        track for track in tracks if track.source_name.lower() == source_name.lower() and track.genre_name == genre
-    ]
-    if not filtered_tracks:
-        logger.info(f"No tracks found for {source_name} in genre {genre}")
-        return
+    def merge_track_data(self, existing_data: dict, new_data: dict) -> dict:
+        def recursive_merge(d1: dict, d2: dict) -> dict:
+            for key, value in d2.items():
+                if isinstance(value, dict) and key in d1 and isinstance(d1[key], dict):
+                    d1[key] = recursive_merge(d1[key], value)
+                else:
+                    d1[key] = value
+            return d1
 
-    sorted_tracks = sorted(filtered_tracks, key=lambda x: x.rank)
-    logger.info(f"Filtered and sorted {len(filtered_tracks)} tracks for {source_name} in genre {genre}")
-    playlist_url = SERVICE_CONFIGS[source_name]["links"][genre]
-    document = {
-        "service_name": source_name,
-        "genre_name": genre,
-        "playlist_url": playlist_url,
-        "tracks": [track.to_dict() for track in sorted_tracks],
-    }
-    insert_or_update_data_to_mongo(mongo_client, TRANSFORMED_DATA_COLLECTION, document)
-    logger.info(f"Inserted {len(filtered_tracks)} tracks for {source_name} in genre {genre}")
+        return recursive_merge(existing_data.copy(), new_data)
 
+    def overwrite(self) -> None:
+        logger.info("Saving transformed data to MongoDB")
+        formatted_tracks = [
+            {"isrc": isrc, "track_data": track.model_dump()}
+            for isrc, track in self.tracks.items()
+            if track.isrc is not None
+        ]
+        formatted_ranks = self.format_playlist_ranks()
 
-if __name__ == "__main__":
-    logger.info("Starting transform")
-    set_secrets()
-    mongo_client = get_mongo_client()
-    logger.info("Clearing transformed playlists collection")
-    clear_collection(mongo_client, TRANSFORMED_DATA_COLLECTION)
-    transform_playlists(mongo_client)
-    logger.info("Transform completed")
+        track_collection = self.mongo_client.get_collection(TRACK_COLLECTION)
+        for track in formatted_tracks:
+            existing_track = track_collection.find_one({"isrc": track["isrc"]})
+            if existing_track:
+                merged_track = self.merge_track_data(existing_track, track)
+                track_collection.update_one({"isrc": track["isrc"]}, {"$set": merged_track}, upsert=True)
+            else:
+                track_collection.update_one({"isrc": track["isrc"]}, {"$set": track}, upsert=True)
+
+        self.mongo_client.overwrite_collection(TRACK_PLAYLIST_COLLECTION, formatted_ranks)
